add channel icon url
