+++ discord/channel.go
@@ -1,5 +1,9 @@
 package discord
 
+import (
+	"github.com/DisgoOrg/disgo/rest/route"
+)
+
 // ChannelType for interacting with discord's channels
 type ChannelType int
 
@@ -92,4 +96,11 @@ type PartialChannel struct {
 	ID   Snowflake   `json:"id"`
 	Type ChannelType `json:"type"`
 	Name string      `json:"name"`
+	Icon *string     `json:"icon,omitempty"`
+}
+
+// GetIconURL returns the Icon URL of this channel.
+// This will be nil for every discord.ChannelType except discord.ChannelTypeGroupDM
+func (c *PartialChannel) GetIconURL(size int) *string {
+	return FormatAssetURL(route.ChannelIcon, c.ID, c.Icon, size)
 }
